+++ zapisy/apps/theses/assets/components/Dialogs/dialog_styles.less
